pyproject.toml
# pyproject.toml

[project]
name = "sandtable"
version = "0.2.0"
description = "Event-driven backtesting and research simulation platform for systematic trading strategies"
readme = "README.md"
requires-python = ">=3.13.1,<3.14"
license = {file = "LICENSE"}

authors = [{ name = "Westimator" }]
dependencies = [
    "ipykernel>=6.29",
    "matplotlib>=3.8",
    "mysql-connector-python>=9.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "plotly>=5.18",
    "pyarrow>=14.0",
    "pytest>=8.0",
    "pytest-cov>=7.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "ruff>=0.4",
    "scipy>=1.11",
    "streamlit>=1.30",
    "tabulate>=0.9",
    "weasyprint>=62.0",
    "yfinance>=0.2",
]

[project.urls]
Repository = "https://github.com/westimator/sandtable"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sandtable"]

[tool.ruff]
line-length = 125
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=sandtable"
